fix : 問題のダウンロード処理をproblemに分離
